+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Dates are in YYYY-MM-DD format.
 - Leverages `SLog`'s `operator<<` overloads to enable displaying better errors.
 - `EXPECT_*` macros now throw a `TestFailedException` to indicate that a failure occurred. This allows them to be used in helper functions as well.
 - Adds `STEST_MAIN` macro to save on boiler plate.
+- Enables tests with duplicate names (must be in different namespaces).
 
 ## v0.2.0 (2019-07-15)
 - Adds `EXPECT_THROWS` and `EXPECT_THROWS_ANY` to check for exceptions.